Added eslint rule, added logging model, refactored login/signup
