+++ my-app/react-app/.eslintrc.json
@@ -42,6 +42,7 @@
       { "extensions": [".js", ".jsx", ".ts", ".tsx"] }
     ],
     "react-hooks/rules-of-hooks": "error",
-    "react-hooks/exhaustive-deps": "warn"
+    "react-hooks/exhaustive-deps": "warn",
+    "react/require-default-props": "off"
   }
 }
