+++ my-app/react-app/src/components/login/models/LoggingModel.ts
@@ -1,0 +1,4 @@
+export interface LoggingModel {
+  email: string
+  password: string
+}
